+++ template/tests/Hypermedia/WorkflowTest.php
@@ -4,18 +4,15 @@ declare(strict_types=1);
 
 namespace BEAR\Skeleton\Hypermedia;
 
-use BEAR\Dev\Http\WorkflowTestTrait;
+use BEAR\Dev\Http\AbstractWorkflowTest;
 use BEAR\Resource\ResourceInterface;
 use BEAR\Resource\ResourceObject;
 use BEAR\Skeleton\Injector;
-use PHPUnit\Framework\TestCase;
 
 use function assert;
 
-class WorkflowTest extends TestCase
+class WorkflowTest extends AbstractWorkflowTest
 {
-    use WorkflowTestTrait;
-
     protected function newResource(): ResourceInterface
     {
         $resource = Injector::getInstance('app')->getInstance(ResourceInterface::class);